requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
requests>=2.23.0
opencv_python_headless>=4.2.0.34
scipy>=1.5.0
scikit_image>=0.17.2
numpy>=1.18.0
Pillow>=5.4.0
skimage>=0.16.1
torchmetric>=1.4.0